Worker TSA registration with unique names
Generate a GUID in powershell to use for the worker name when the
instance registers with the concourse TSA. Without a unique name, the
default is to use the instance hostname.

On the windows AWS stemcell, the instances do not have unique hostnames
after creation. Multiple workers attempting to register with the same
hostname used as the worker name has "unexpected behavior".

Signed-off-by: Jim Doty <jdoty@pivotal.io>